+++ lib/localize.dart
@@ -5,9 +5,11 @@
 library;
 
 export 'src/localize/client.dart';
+export 'src/localize/google_sheets.dart';
 export 'src/localize/language_names.dart';
 export 'src/localize/localizer.dart';
 export 'src/localize/models.dart';
 export 'src/localize/prompt.dart';
+export 'src/localize/sheets.dart';
 export 'src/localize/utils.dart';
 export 'src/localize/validation.dart';